types des variables => (u)int(8|16|32)_t - /!\ x et y doivent être dans l'écran drawDisc.c corrigé -> le disque était décalé

+++ src/nRGBlib/drawBox.c
@@ -2,9 +2,9 @@
 #include <os.h>
 
 /// Dessine un rectangle en RGB
-void drawBox_(short x, short y, short w, short h, Color c)
+void drawBox_(int16_t x, int16_t y, int16_t w, int16_t h, Color c)
 {
-    short i, j;
+    int32_t i, j;
     for (i = x; i < (x + w) && i < 320; i++)
     {
         if (i < 0) continue;
@@ -20,7 +20,7 @@ inline void drawBox(Box *box, Color c)
     drawBox_(box->x, box->y, box->w, box->h, c);
 }
 
-inline void drawSquare_(short x, short y, short side, Color c)
+inline void drawSquare_(int16_t x, int16_t y, int16_t side, Color c)
 {
     drawBox_(x, y, side, side, c);
 }
